Make CLI loop the current program after default error handler.
This is mostly to address the case where the host has sent code which is
unsuitable for the tube processor, and consequently not executed, but
which has overwritten the current program.  Previously this client would
enter the CLI loop but not update HIMEM or the current program pointer
so upon break would try to re-execute the overwritten old current
program and crash.

Making the CLI loop the current program when the default error handler
is invoked solves this but a futher test is required because HIMEM
cannot be the address of the CLI loop but needs to be the address of the
start of the ROM.